% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{epa}
\alias{epa}
\title{Datatable of the epa Example}
\format{
A data frame with 620 observations in 6 variables
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(epa)
}
\description{
Datatable of the epa Example
}
\details{
Variables in the dataset:
\itemize{
  \item sex. a factor with levels \code{m} \code{w} \code{d}, giving the proband's sex
  \item age. a numeric vector
  \item cms. a numeric vector
  \item risk. a dichotome vector, 0 = not at risk, 1 = at risk
  \item expert. a dichotome vector of expert's decision, 0 = not at risk, 1 = at risk
  \item decu. a dichotome vector, 0 = no decubitus, 1 = decubitus
}
}
\keyword{datasets}
