% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Faktorenbogen}
\alias{Faktorenbogen}
\title{Datatable of the Faktorenbogen Example for factor analysis}
\format{
A data frame with 150 observations in 14 variables
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(Faktorenbogen)
}
\description{
Datatable of the Faktorenbogen Example for factor analysis
}
\details{
Variables in the dataset:
\itemize{
  \item gender. a factor with levels \code{female} \code{male} \code{other}, giving the proband's gender
  \item age. a numeric vector of proband's age in years
  \item A. Item A of the questionnaire, numeric
  \item B. Item B of the questionnaire, numeric
  \item C. Item C of the questionnaire, numeric
  \item D. Item D of the questionnaire, numeric
  \item E. Item E of the questionnaire, numeric
  \item F. Item F of the questionnaire, numeric
  \item G. Item G of the questionnaire, numeric
  \item H. Item H of the questionnaire, numeric
  \item I. Item I of the questionnaire, numeric
  \item J. Item J of the questionnaire, numeric
  \item K. Item K of the questionnaire, numeric
  \item L. Item L of the questionnaire, numeric
}
}
\keyword{datasets}
