% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{freqTable}
\alias{freqTable}
\title{create a frequency table}
\usage{
freqTable(werte)
}
\arguments{
\item{werte}{factor with obeserved data}
}
\value{
dataframe table
}
\description{
returns a frequency table with absolute and relative frequencies and cumulated frequencies
}
\examples{
x <- ceiling(stats::rnorm(20))
freqTable(x)

}
