% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kenngroessen}
\alias{kenngroessen}
\title{create a tibble with kenngroessen}
\usage{
kenngroessen(werte)
}
\arguments{
\item{werte}{numeric vector}
}
\value{
tibble with all kenngroessen
}
\description{
returns a tibble with all kenngroessen
}
\examples{
x <- ceiling(stats::rnorm(20))
kenngroessen(x)

}
