% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KIbinomail_a}
\alias{KIbinomail_a}
\title{compute confidence intervall  for binomial proportions}
\usage{
KIbinomail_a(p, n, alpha)
}
\arguments{
\item{p}{proportion obeserved}

\item{n}{number of observations}

\item{alpha}{error niveau}
}
\value{
confidence intervall
}
\description{
returns borders and length of confidence intervall for binomial proportions
}
\examples{
KIbinomail_a(0.35, 150, 0.05)

}
