% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KIbinomial_u}
\alias{KIbinomial_u}
\title{compute confidence intervall for difference of binomial proportions}
\usage{
KIbinomial_u(p1, n1, p2, n2, alpha)
}
\arguments{
\item{p1}{proportion obeserved in group 1}

\item{n1}{number of observations in group 1}

\item{p2}{proportion obeserved in group 2}

\item{n2}{number of observations in group 2}

\item{alpha}{error niveau}
}
\value{
confidence intervall
}
\description{
returns borders and length of confidence intervall for difference of binomial proportions
}
\examples{
KIbinomial_u(0.25, 100, 0.4, 150, 0.05)

}
