% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KInormal_a}
\alias{KInormal_a}
\title{compute confidence intervall  for mean of normal distributed data}
\usage{
KInormal_a(xquer, s, n, alpha)
}
\arguments{
\item{xquer}{mean of obeserved data}

\item{s}{standard deviation of observed data}

\item{n}{number of observations}

\item{alpha}{error niveau}
}
\value{
confidence intervall
}
\description{
returns borders and length of confidence intervall for mean of normal distributed data
}
\examples{
KInormal_a(400, 20, 100, 0.05)

}
