% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KInormal_u}
\alias{KInormal_u}
\title{compute confidence intervall  for mean of normal distributed data}
\usage{
KInormal_u(x1, s1, n1, x2, s2, n2, alpha)
}
\arguments{
\item{x1}{mean of obeserved data in group 1}

\item{s1}{standard deviation of observed data in group 1}

\item{n1}{number of observations in group 1}

\item{x2}{mean of obeserved data in group 2}

\item{s2}{standard deviation of observed data in group 2}

\item{n2}{number of observations in group 2}

\item{alpha}{error niveau}
}
\value{
confidence intervall
}
\description{
returns borders and length of confidence intervall for mean of normal distributed data
}
\examples{
KInormal_u(2.22, 0.255, 13, 2.7, 0.306, 10 , 0.05)

}
