% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lon.lat.osm}
\alias{lon.lat.osm}
\title{get longitude and altitude from an address
using OpenStreetMap's API at
http://nominatim.openstreetmap.org}
\usage{
lon.lat.osm(address = NULL)
}
\arguments{
\item{address}{a character of an address}
}
\value{
a data.frame containig "address", "lon", "lat"
}
\description{
get longitude and altitude from an address
using OpenStreetMap's API at
http://nominatim.openstreetmap.org
}
\examples{
lon.lat.osm("Eiffeltower")

}
