% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MarioANOVA}
\alias{MarioANOVA}
\title{Datatable of the SuperMario Example for Friedman-ANOVA}
\format{
A data frame with 47 observations in 8 variables
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(MarioANOVA)
}
\description{
Datatable of the SuperMario Example for Friedman-ANOVA
}
\details{
Variables in the dataset:
\itemize{
  \item Name. The characters' name
  \item Alter. The characters' age in years
  \item Kingdom. The characters' home
  \item Geschlecht. The characters' gender (männlich = male, weiblich = female)
  \item BadGuy. Whether the character is a bad guy, logical
  \item t1. Measure at time 1
  \item t2. Measure at time 2
  \item t3. Measure at time 3
}
}
\keyword{datasets}
