% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Messwiederholung}
\alias{Messwiederholung}
\title{Datatable of the Messwiederholung Example for ANOVA}
\format{
A data frame with 200 observations in 4 variables
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(Messwiederholung)
}
\description{
Datatable of the Messwiederholung Example for ANOVA
}
\details{
Variables in the dataset:
\itemize{
  \item Name. The first name of the probands.
  \item t1. Measure at time 1
  \item t2. Measure at time 2
  \item t3. Measure at time 3
}
}
\keyword{datasets}
