% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mma}
\alias{mma}
\title{Dataset of a work sampling study}
\format{
A data frame with 9768 observations in 6 variables.
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(mma)
}
\description{
Dataset of a work sampling study
}
\details{
Variables in the dataset:
\itemize{
  \item day. a vector, giving the number of the observation day
  \item time. a factor giving the time of observation
  \item ward. a factor giving the ward under observation
  \item qual. a factor giving the qualification of the nurse
  \item category. a factor of qualification categories
  \item action. a factor giving the observed action
}
}
\keyword{datasets}
