% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Nachtwachen}
\alias{Nachtwachen}
\title{Dataset of the German Nachtwachen study}
\format{
A data frame with 276 observations in 37 variables.
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(Nachtwachen)
}
\description{
Dataset of the German Nachtwachen study
}
\keyword{datasets}
