% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nw}
\alias{nw}
\alias{nw_labelled}
\title{Dataset of the German Nachtwachen study with labelled variables}
\format{
A data frame with 276 observations in 37 variables.
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(nw)
}
\description{
Dataset of the German Nachtwachen study, labelled version
}
\keyword{datasets}
