% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OrdinalSample}
\alias{OrdinalSample}
\alias{ordinalSample}
\title{Datatable of an Ordinal Sample}
\format{
A data frame with 415 observations in 4 variables.
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(OrdinalSample)
}
\description{
Datatable of an Ordinal Sample
}
\details{
Variables in the dataset:
\itemize{
  \item Konflikt. a numeric vector giving the potential of conflicts.
  \item Zufriedenh. a numeric vector giving the satisfaction of workers
  \item Geschlecht. a factor of proband's sex, 1 = male, 2=female
  \item Stimmung.  an ordinal factor of proband's mood
}
}
\keyword{datasets}
