% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pf8}
\alias{pf8}
\title{Dataset of the PF8 example.}
\format{
A data frame with 731 observations in 16 variables.
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(pf8)
}
\description{
This is the dataset of the PF8 example.
}
\keyword{datasets}
