% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Pflegeberufe}
\alias{Pflegeberufe}
\title{Matrix of Pflegeberufe by Isfort et al. 2018}
\format{
A matrix with 9 cols (years) and 5 rows (nursing profession).
}
\source{
\url{https://www.produnis.de/R/}
}
\usage{
data(Pflegeberufe)
}
\description{
Matrix of Pflegeberufe by Isfort et al. 2018
}
\author{
Isfort et al. 2018 (Pflegethermometer)
}
\keyword{datasets}
