% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sens.spec}
\alias{sens.spec}
\title{compute sensitivity and specifity}
\usage{
sens.spec(rp, rn, fp, fn)
}
\arguments{
\item{rp}{number of true-positive (richtig-positiv)}

\item{rn}{number of true-negative (richtig-negativ)}

\item{fp}{number of false-positive (falsch-positiv)}

\item{fn}{number of false-negative (falsch-negativ)}
}
\value{
a data.frame with sens, spec, ppw, npw
}
\description{
returns sensitivity specifity, negativ-predictive-value, postitiv-predictive-value
}
\examples{
sens.spec(40, 17, 85, 4)

}
