% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ztrans}
\alias{ztrans}
\title{z-Transformation by given numbers, with
z = (x - mu) / sd}
\usage{
ztrans(x, mu = 0, sd = 1)
}
\arguments{
\item{x}{a value to transform}

\item{mu}{the given mu}

\item{sd}{the given standard deviation}
}
\value{
the z-transformed value
}
\description{
z-Transformation by given numbers, with
z = (x - mu) / sd
}
\examples{
ztrans(120,mu=118,sd=20)

}
